+++ NEWS.md
@@ -1,3 +1,12 @@
+# softermax NEXT
+
+* `as.data.frame` functions: column order change in resulting data frames
+* `as.data.frame` functions: explicit `row.names` and `optional` parameters for compatability with generic
+* Added `experiment_names` and `plate_names` functions to extract experiment names and plate names
+* `read_softmax6_xml` now warns if experiment contains multiple plates with the same name
+* Added `as_tibble` methods to coerce data to [tibbles](http://tibble.tidyverse.org) (with optional nesting)
+
+
 # softermax 0.2.0
 
 * `read_softmax_xml` now returns a softermax object, which maintains more of the original data and metadata
@@ -10,6 +19,3 @@
 # softermax 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
-
-
-