+++ NEWS.md
@@ -1,4 +1,4 @@
-# softermax NEXT
+# softermax 0.2.2
 
 * Added `apply_template` to annotate read data with experiment information stored in plate templates
 * Added ReadMode column to `as.data.frame` to better support multi-mode experiments